[Feat]: 그룹원 추가 페이지 구현
- App.tsx 너비 제한 360px
- InputField.tsx로 작성 폼 구현
- SelectField.tsx로 드롭다운 구현
- 재사용 가능한 컴포넌트를 common으로 분리
- mui 대신 tailwind 사용
